1.0.0.5 -Removed renaming of cache.psarc in SetlistManager -Highlight dgvDlcSongs that have already been used in a setlist -Removed some redundant code from SetlistManager, more code optimization -Added 'Delete Selected' button to SongManager with saftety interlock checkbox -Fixed SetlistManager UC window docking and positioning when in release mode -Fixed NoteViewer window position to always be topmost in release mode

+++ CustomsForgeManager/Program.cs
@@ -22,7 +22,7 @@ namespace CustomsForgeManager
                 if (Process.GetProcessesByName(Process.GetCurrentProcess().ProcessName).Length > 1)
                     return;
 
-            if (FirstRun())
+            if (FirstRun() && !Constants.DebugMode)
             {
                 if (Directory.Exists(Constants.WorkDirectory))
                     Directory.Delete(Constants.WorkDirectory, true);
